feat: SHA256 Checksum 驗證 + record_duration 改為整數秒
- minio_service: create_multipart_upload 加入 ChecksumAlgorithm="SHA256"；generate_part_upload_url Params 加入 ChecksumAlgorithm；complete_multipart_upload 傳入含 ChecksumSHA256 的 parts
- upload_job_service: complete_multipart 組裝 ChecksumSHA256 per part
- schemas/upload_job: MultipartInitResponse 新增 checksum_algorithm；PartCompleteRequest 新增 checksum_sha256
- models/audio + schemas/audio: record_duration 型別由 float 改為 int（秒整數）
- alembic: 新增 merge heads migration 與 change_record_duration_to_integer migration
- entrypoint.sh: alembic upgrade head → heads（支援多 head）
- tests: 整合測試加入 x-amz-checksum-sha256 header 驗證
- 移除 query_test/ 目錄（開發期臨時查詢腳本）
- 新增 scripts/upload_test.py、upload_test_cleanup.py
- 更新 .claude 文件（plans/batch-audio-upload.md、docs/batch-audio-upload.md）

+++ app/schemas/audio.py
@@ -21,7 +21,7 @@ class AudioBase(BaseModel):
     file_size: int | None = Field(None, gt=0, le=10_737_418_240)
     checksum: str | None = None
     record_time: datetime | None = None
-    record_duration: float | None = Field(None, gt=0)
+    record_duration: int | None = Field(None, gt=0)
     fs: int | None = Field(None, ge=0, le=384000)
     recorder_channel: int | None = Field(default=0, ge=0, le=16)
     audio_channels: int | None = Field(default=1, ge=1, le=16)
@@ -52,7 +52,7 @@ class AudioUpdate(BaseModel):
     file_size: int | None = None
     checksum: str | None = None
     record_time: datetime | None = None
-    record_duration: float | None = None
+    record_duration: int | None = None
     fs: int | None = Field(None, ge=0, le=384000)
     recorder_channel: int | None = None
     audio_channels: int | None = None
@@ -128,7 +128,7 @@ class AudioBatchItem(BaseModel):
     file_size: int | None = Field(None, gt=0, le=10_737_418_240)
     checksum: str | None = None
     record_time: datetime | None = None
-    record_duration: float | None = Field(None, gt=0)
+    record_duration: int | None = Field(None, gt=0)
     fs: int | None = Field(None, ge=0, le=384000)
     recorder_channel: int | None = Field(default=0, ge=0, le=16)
     audio_channels: int | None = Field(default=1, ge=1, le=16)
